Convert LightyServerBuilder to LightyJettyServerProvider
This conversion is required in order to use JaxRsEndpoint and
JaxRsOpenApi since they do not use regular server but JettyWebServer.

This patch basically replaces LightyServerBuilder with JettyServerProvider
which makes JettyServer acessible everywhere in Lighty.io where it may be
needed.

This code contains TODO to convert HTTP2 and HTTPS to JettyWebServer
which should be done in separate task due to scale of LIGHTY-329.

JIRA: LIGHTY-329
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>